iter1 Правка типа

+++ migrations/000002_create_orders_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE orders (
     user_id UUID REFERENCES users(id) ON DELETE CASCADE,
     order_id VARCHAR(255) NOT NULL UNIQUE,
     status VARCHAR(50) NOT NULL,
-    points_accumulated INT DEFAULT 0,
+    points_accumulated FLOAT DEFAULT 0,
     withdrawal_status VARCHAR(50) DEFAULT 'PENDING',
     created_at TIMESTAMP DEFAULT NOW()
 );
